core/model: Allow & use non-critical cross-thread exceptions.
This commit adds a new boolean `critical` parameter to
raise_exception_in_main_thread. If the exception event is not marked as
critical then a notice popup window is shown prompting users to report
the issue to #zulip-terminal or to the github issue page. If it is
critical, the previous behavior is retained, ie. the exception from the
other thread is raised in the main thread and the application exits.

With this new parameter set to `False`, the poll_for_events loop can
fail for one event (with a notice) but continue otherwise to allow the
application to maintain more functionality without either abruptly
exiting or not indicating that something has subtly failed.