requirements/tools: Upgrade isort from ==4.3.21 to ~=5.6.0.
With the removal of support for python 3.5, there is no reason not to
upgrade isort beyond 4.3.21. For now we loosely pin to 5.6.*.

With this change the recursive option becomes the default, so the
linting script and makefile are updated accordingly; the feedback text
is also updated.

This change requires a minor change to setup.py, and extends the isort
coverage to all of tools/, so also updates the formatting of another
script.