model: Add function to send typing status.
This commit adds a function to send typing events to the server. The
function works asynchronously to avoid a lag between keypress and the
corresponding letter's display on the screen.

Tests added.

The write_box() fixture function in test_boxes.py currently does not have
a to_write_box attribute which causes test failures for to_write_box
calls inside the keypress tests. This commit sets it to None to support
such calls.