boxes: Configure private_box_view to send typing status.
This commit captures the user's typing status using urwid's
connect_signal(), observing for changes in the msg_write_box. The
msg_write_box is connected to a function that sends 'start' events with
a limit on how frequently the functions can be called. Idleness is
tracked with a method which is initiated when a start event is sent
and checks for a possible halt. The corresponding wait periods after the
user starts and stops typing are used to restrict calling the function
excessively.

Check for empty recipient_user_ids added by neiljp.

Fixes #593.