refactor: boxes: Define and use named focus values within WriteBox.
The expected values of focus_position and focus_col for each box in a
WriteBox class are declared as member variables. This makes it much
clearer to follow the logic when the focus is changed.

This commit also aims to refactor self.contents[0][0] to
self.header_write_box wherever necessary.