refactor: boxes: Focus column/container values centralized.
The focus_position and focus_col values for each box in a WriteBox
class is declared as a member variable. This would help in easily
changing focus value for a given column/container.

This commit also aims to refactor self.contents[0][0] to
self.header_write_box wherever necessary.

Tests refactored and updated (the ones linked with focus values).