boxes: Header wrapped in single line box; Other styling improvements.
Private box and stream box headers both are now Column widgets (for
uniformity). The header boxes are wrapped in a LineBox, so that
bottom edge off-set behavior is fixed. This fixes #629.

The vertical dividing lines in header are removed. The top-line of
header box is made thicker for visual improvement.

The edit mode button is appended as it is (i.e., without wrapping in a
seperate LineBox).