boxes/helper: Fixed generation of correct quote fences.
This commit is a bug-fix that corrects generation of fences for quoted
messages. The logic for generation of quote fence is borrowed from the
webapp. Fences are generated after searching for maximum length occurence
of ` through regex, and wrapping the quote block with the generated fence.
This allows for proper quoting of messages containing a code-block or
another quote inside them.
This is the first part of commit for this Pull-request.
Tests amended.