bugfix: boxes/helper: Fix generation of correct quote fences.
This commit corrects generation of fences for quoted messages. The new
logic for generation of quote fence is borrowed from the webapp. Fences
are generated after searching for maximum length occurence of ` through
regex, and wrapping the quote block with the generated fence. This
allows for proper quoting of messages containing a code-block or another
quote inside them.

Tests added.

Math & spoiler cases, and test ids, added by neiljp.