boxes/server_url/docs: Update to use new webapp quoting style.
This commit adds the same new quoting style as used in the webapp, ie.
indicating the quoted author and location.

The URL generation part of the code now has its own module named
server_url.py, along with the tests for this module added as
test_server_url.py. This will allow to track changes in the server and
make constructing and deconstructing server URL's easier. This module
can also be later integrated into the API package.

Details of server_url.py are added to docs/developer-file-overview.md

Tests added and amended.

Fixes #514.