boxes/server_url: Updated to new quoting style.
This is the second commit to this PR, which adds the same
quoting style as used in the webapp. The URL generation
part of the code now has its own module named server_url.py
along with the tests for this module added as test_server_url.py.
This will allow to track changes in the server and make constructing
and deconstructing server URL's easier. This module can also be later
integrated into the API package.

Tests amended.
Fixes #514.