views: Added stream message weekly count.
This is the second commit of this PR, which adds the
option of weekly message count for streams from the
Stream info view. The stream info is now updated to
contain the option to view the list of members
subscribed to a stream along with the weekly count of
messages in stream. These statistics help get a better
overview of the streams and acts as a feature enhancement.

Tests amended.