refactor: buttons: Pass user_id to private_box_view in UserButton.
Currently, the UserButton only passes itself as the `button` parameter
to the private_box_view() when pressed, on the users list. This leads to
an inconsistency as the corresponding user_id isn't passed. This commit
passes the user_id explicitly, to ensure availability of parameters
should there be any need inside the private_box_view in the future.