helper/model: Add footer notification for message sent outside narrow.
This commit generalizes the code for footer notification when a message
is sent/edited outside the current narrow. This will give the user a pointer
as to why the message disappeared from the screen. The code is hooked to
the part where message request is sent to the server, hence preventing the
extra check for user_id being same as message_sender_id.

Tests amended.

Fixes #781.