model/views: Add stream creation date in Stream Info popup.
This commit adds stream creation date inside StreamInfo view that
uses model.preferred_local_time to display date and time in an
extended format including the year.

The date_created field was added in Server version 4.0, feature
level 30, which means this field would only be present in server
iterations from this. To maintain consistency, we add this field
in model's stream data even for server iterations before this with
a value of None.

Tests amended.