tools: Use and apply isort with multi-line style '3' for smaller diffs.
This causes `from` imports between parentheses to be placed on separate
lines. This makes them easier to read, and additions/removals/updates
only affect one line rather than shifting text between lines when they
re-wrap.

This also will enable compatibility with black, if we move to that tool
in the future.