bugfix: model: Fix incorrect condition in update_message_event.
This commit fixes an incorrect condition in handle_update_message_event
that previously assumed subject key would be present in event, only
when the message's topic/subject was edited. This has been recently
proven incorrect, because of the fact that subject is always present
in event if the message is edited.

Tests amended.