refactor: model/boxes/views: Centralize function for time formatting.
This commit replaces the _time_for_message() and ctime() methods
in boxes and views with the centralized method formatted_local_time()
in model.Currently this function processes the message timestamp and
returns the default configuration of date and time.
The functionality can be extended to support modification of date and
time according to user settings in future commits.

Tests modified.