refactor: boxes: Show recipient names along with emails in PMs.
Currently, PM narrows only display the delivery emails of the users. This
commit modifies the to_write_box to also display the names of the users
along with their emails in the 'name <email>' format. Only the emails
are validated while cycling out of the to_write_box.

Tests updated.