model/views: Use consistent last-message tracking to avoid API call.
Prior to this commit, pressing GO_DOWN at the bottom of a message view
in which there were no more messages would hit the API each time,
looking for newer messages.

With the consistent new-message tracking now present in the model, a new
model accessor then supports determining this state, avoiding scrolling
and fetching newer messages if not required.