model/views: Add first-message tracking per narrow to avoid API calls.
Prior to this commit, pressing GO_UP at the top of a message view in
which there were no more messages would hit the API each time, looking
for older messages.

If the oldest message in a narrow is reached, this is now tracked in the
model much like tracking the last message (the API uses
found_[oldest|newest]). A new model accessor then supports determining
this state, avoiding scrolling and fetching older messages if not
required.