model/buttons/views: Event handler for muted_topic event for (un)muting.
We register the event_type 'muted_topics' to the event queue on startup,
along with a handler method which responds to these events appropriately.
Depending on whether the topic is already muted or not, the handler method
calls the `mark_muted` or `mark_unmuted`. These two functions will update
the data structures which keeps track of message counts in a later commit.

Tests added