boxes: Remove space to split recipients by ','.
Prior to this commit, the text in the to_write_box used by the
autocomplete was split by ', ' which made it mandatory for a space
to be included after ','. This commit splits the recipients in the
to_write_box by ',', making it more flexible in terms of autocompletion.

Tests updated.