refactor: core/boxes/buttons: Rename & pass args narrowing to all msgs.
This replaces the previous behavior of passing an object with parameters
expected to be present (eg. a button), with simply optionally the
message id.

While the view is not "narrow", the method name is updated for
consistency to start with "narrow_to_".

Since the HomeButton provides no parameters, there is no longer a need
for functools.partial in this case.

Tests updated.