refactor: core/boxes/buttons/ui: Rename & pass args narrowing to all PM.
This replaces the previous behavior of passing an object with parameters
expected to be present (eg. a button), with simply optionally the
message id.

The method name is updated for consistency to start with "narrow_to_".

Since the PMButton provides no parameters, there is no longer a need
for functools.partial in this case.

Tests updated.