refactor: core/ui/buttons: Rename & be explicit narrowing to starred.
No calling point passes a message id, so all parameters to the method to
narrow to starred messages in the Controller are removed.

The method name is also updated for consistency to start with
"narrow_to_".

Since the StarredButton provides no parameters, there is no longer a
need for functools.partial in this case.

Tests updated.