refactor: core/ui/buttons: Rename & be explicit narrowing to mentions.
No calling point passes a message id, so the parameter to the method to
narrow to mentions in the Controller is removed.

The method name is also updated for consistency to start with
"narrow_to_".

Since the MentionedButton provides no parameters, there is no longer a
need for functools.partial in this case.

Tests updated.