boxes: Use `user_id`s in autocomplete for ambiguous user mentions.
Prior to this commit, user mentions through autocomplete used only
names in the typeahead. This commit appends `user_id`s to all the
multiple instances of the same name to use the **Name|User_id** format of
user mentions.

This is done by using collections.Counter to check the count of each user
name in the users list. The user_id is appended to names that have a count
greater than 1. The appended user_id would serve as a distinction
between users having the same name.

Tests added and updated.

Fixes #151.