refactor: boxes: Extract footer notification for invalid stream/email.
This commit refactors code for footer notification in case of invalid
recipients (stream/email) in the WriteBox to separate helper function.
This prevents duplication of code and allows reusability.

This function can now be called with a custom error message that will
be displayed in the footer after appending it with a common typeahead
suggestion.

Tests amended.