buttons/views: Display and sort emojis by count in EmojiPicker view.
A new reaction_count field is added to the EmojiButton class to keep
in check the number of reactions with a particular emoji on a message.

The EmojiPicker view also shows the count of reactions of a particular
emoji beside emoji_name, and the list is sorted in descending order of
emoji_count, then in ascending order of emoji_names if the count is equal.

Tests amended.