bugfix: keys/hotkeys: Revert minor changes to improve documentation.
A few changes in fb9f20b56 had no apparent issue with UI navigation, but
unintended consequences with regard to triggering custom actions. This
commit therefore reverts a subset of that commit to allow those to
proceed as before.

In particular, scrolling beyond the current list of messages was not
possible with the renamed keys, as the generated key was 'down' but the
listed entry was 'Down'. Similarly with other directions, page up/down,
and End.

docs/hotkeys.md regenerated.