bugfix: boxes: Improve quoted text rendering, fixing previous xfails.
This commit fixes some of the cases of incorrectly rendered quote texts.
This includes nested quotes and multiline quotes up to at least the
second level. This is a partial bugfix, as there are other cases of
incorrect rendering of quotes; extra xfails are added to cover some of
these cases.

Tests amended and added.

Partial fix towards #402.