bugfix: model: Update messages in starred narrow on message flag events.
Messages were individually updated as (un)starred upon events, but the
central record of all starred messages was not. The latter is used to
identify messages within the starred messages narrow, so the record of
starred messages could not previously change during a session.

Like in the webapp, this does not instantly remove messages from the
view upon unstarring them, but updates the model so that a return to the
starred messages narrow will do so.

Tests updated.

Fixes #940.