ui/boxes: Improve consistency of non-reply compose key behaviour.
Previously, `x` started a new message as expected from all locations,
with no pre-filled recipient.
In contrast, `c` would act as per the reply key on a private message,
would pre-fill the private recipients or stream, and was configured for
use only when focus was on the `MiddleColumnView`.

Now keypresses `x` and `c` always open up `private_box_view` and
`stream_box_view` respectively, with no boxes pre-filled. The scope of
`c` is expanded to enable its use even when focus is on the users list,
stream list, topics list, or top buttons.

Note that this behavior is slightly different from the webapp, where in
a stream narrow, `c` pre-fills the stream and prompts for a topic. This
is a slight change in behavior to before within the application too.

Tests added, and adjusted slightly by neiljp.

Fixes #871.