bugfix: boxes/views: Make 'r' independent of 'enter' for message_reply.
The 'r' key that is used to reply to a message was actually depending
on 'enter' key for its functionality. When replying to a message, the
keypress is first intercepted by MiddleColumnView as part of REPLY_MESSAGE
command, which is passed to the corresponding MessageBox as "enter"
keypress. The MessageBox grabs this key and executes message reply but via
the ENTER filter.

This commit makes 'r' key for replying to message independent of 'enter' by
passing the command 'REPLY_MESSAGE' instead of the hardcoded 'enter' key to
MessageBox.

Tests amended.

Balance of code between commits adjusted slightly for correctness by neiljp.