setup/scripts: Add render_symbols to enable testing of symbol rendering.
This commit adds a new folder `scripts/` to ship scripts that are meant
for users. This commit also adds a new script that enables users to check
how the various symbols listed in `config/symbols.py` render on their
terminal emulator.

The need for this script has arisen from the observed issues on the Windows
Terminal, where with varying fonts, symbol rendering has also varied.