run: Add --notify and no-notify options as command-line arguments.
This commit adds new command-line arguments for notify and no-notify
that corresponds to desktop notifications being enabled or disabled
in the current session, respectively. This should allow the user to
specify a notify option and override its corresponding value in the
zuliprc.

This commit is a potential prep for #900, that enables modifying
Stream notification settings in a running session.

Tests added.