bugfix: core: Fix message not focused within same narrow if anchored.
This is bugfix that allow focusing on messages from the same narrow
if a contextual message-id (anchor) is provided to the _narrow_to function.

Earlier, if the user was within the same topic narrow then narrowing to a
quoted message from MsgLinkButton did not select the message, since the
_narrow_to() function didn't took anchor into account in this case.

Tests amended.

Fixes #954.