view: Fix focus before and after search.
This commit corrects what the code for focus_index_before_search was
intended to do. The focus index was still being updated even in the
search results.

This is fixed by adding a boolean `in_search_mode` which does not
call `get_focus` untill the search is over. Slight reordering was
done in `__init__` to fit the boolean in an ideal place w.r.t.
readability.

Fixes #975