bugfix/feature: views/boxes/themes: Fix crash on no-result search.
This commit fixes a bug which causes a crash when searching
Streams or Topics. This happens because when a search result is
empty, log is empty and because an empty ListWalker returns None,
it leads to a Type error when setting focus.

This is fixed by making sure the user does not enter the empty list
when the search result is empty. The boolean `empty_search` is used
to achieve this.

This commit also improves the UI/UX when search result is empty by
displaying Text `Not Found` in place of the empty list.
The user can then correct their search and `Not Found` disappears.

A `search_error` attribute is introduced in the palette which
colors `Not Found` in `light red`. The `cross_mark` symbol -
`INVALID_MARKER` is also added for effect.

Fixes #923.

Tests amended (slightly improved by neiljp).