refactor: urwid_types/ui: Migrate View.keypress size arg to new type.
The type of size has been changed from `Tuple[int, int]` to a new
`urwid_Box` type. Since the outermost widget is always a Box widget,
which is also true because `Frame` is a Box widget, an `urwid_Box` type
is defined along with `urwid_Flow` and `urwid_Fixed`, which combined
make up the existing `urwid_Size` type.