boxes: Add status marker as a prefix to sender name in content header.
This commit adds a status marker right before the sender's name in a
message content header. The users' dict from model is used to get
status of the sender. The default status of sender is kept `inactive`.

NOTE: Currently bots are marked as `inactive`. A later version would
be to render markers for bots as well.

Tests amended (Message content header related).