ui/views/boxes: Erase `caption` but keep `edit_text` on `q` press.
This commit erases `caption` and keeps `edit_text` while
reattempting a search.

`edit_text` was being erased every time because `search_text` had to
be erased, hence keeping the `edit_text` becomes a problem.

To solve this, the default `search_text` was used in caption instead
and `edit_text` was set to ''. Removing the line that empties text
in ui.py gives the feature we want.

`set_content('')` is used so that now, when a user presses `q` for a
second search the caption becomes empty.

This makes for a cleaner UI while searching by removing caption
`Search Results` when typing.

Tests updated.