refactor: helper: Fixes update errors and unread count edge case.
This is caused when ZT recieves an update event while searching and
`set_count` assumes that all widgets in the log are Buttons since
the log is updated while searching.

This also means that message updates are only sent to a subset of
streams, topics or users. Which also leads to missed updates and
negative counts.

Since, we only need to update the count variable in the button
object this is solved by using `btn_list` instead of `log`.

Solves part of #642