bugfix: helper: Fix unread count inaccuracy from events during searches.
This is caused when an event is received while a search is occurring and
`set_count` previously assumed that the log contained all Buttons rather
than the subset during a search.

This error was present for streams, topic and users.

This is solved by using `btn_list` instead of `log`.

Solves part of #642.