model/helper: Update `set_count` to use index['unread_msgs'].
Use index['unread_msgs'] to update count in `model` and to further
change button count in UI.

Later on when adding support for marking a particular stream or topic
messages as read, index['messages'] may not be effective as it only
stores a few messages. This new data structure may prove useful to have
access to all concerned message ids.

Tests amended.