run: Show warning on startup to enable UTF-8 encoding in locale.
This commit shows the warning message during loading, if UTF-8 encoding
isn't found enabled in the locale.

NOTE: The warning might be shown even if icons are rendering properly. This
might be due to default UTF support for some terminals. However, this could
be a potential risk with other OS (e.g., WSL) and hence robust for such
scenarios.

Test added.

Fixes #256.