FAQ: Remove section suggesting users to set their terminal's encoding.
This commit removes the "Unable to render non-ASCII characters" section
from the FAQ which suggested that users set their terminal's encoding
manually. This is no longer required post the version 0.4.0 release, in
which we introduced urwid's set_encoding() method in run.py to set the
user's terminal's encoding internally to render the characters properly.