FAQ: Update symbol rendering information.
This commit updates the FAQ regarding symbol rendering information
and specifies that the rendering of characters varies based on
emulator and font. It also specifies that we provide a tool for
the user that to conveniently test symbol rendering in their terminal
emulator and font.

A screenshot of the tool run on the GNOME Terminal in Ubuntu is
also displayed so the users can note what the tool is ideally expected
to display.