bugfix: boxes: Set completer delims for message write box.
This small commit explicitly sets delimiter for msg_write_box, in
order to support autocompletion of text with whitespaces, for both
private and stream write boxes. Note that the delims specifed are
default ones, except that whitespace is ignored, so as to make it
function properly in autocompletes. The default completer delims
for message compose are stored in the constant
`DELIMS_MESSAGE_COMPOSE`.

The commit also fixes a bug in which a user isn't able to use
autocompletes for space-containing texts.

Tests amended (comments) and added. The existing tests for
space-containing prefixes are justified with this commit.

Fixes #925 as a side effect.