boxes/ui/model/helper/api_types: Use `user_id`s as message destinations.
This commit migrates from using delivery emails as message destinations
to using `user_id`s instead. The abstraction between the narrow model
and the Composition API type is maintained by modifying the corresponding
connections to the narrow model such that the delivery emails are still
in use in the narrow model currently, making this a message-destination-only
change.

The commit focuses on changing the type of the `recipients` parameter of
the PrivateComposition class to List[int] instead of the previous
List[str]. `send_private_message` now expects a List of `int`s as well.

Tests updated.