helper: Update unread counts for muted topics.
Handles setting count of `model.unread_counts`.
Functions `_set_count_in_model` and `_set_count_in_view` are modified
to keep track of unread counts of muted topics when a new message is
recieved or an existing message is read. This ensures that the count
is in sync when the topic is unmuted later on.