themes: Add pygments style data to themes, including at runtime.
This commit uses Pygments internal dictionaries and PygmentsStyles
to add it into the palette right after theme generation. For this a
META dict is defined in each theme file which contains syntax
highlighting styles and overrides.

The conversion to urwid compatible theme format is done inside
`add_pygments_style` and called in `generate_theme`. The function
does the following:
* Checks if any overrides are necessary.
* Checks if any parent pygment style can be used if it is not defined
* Adds new_style to urwid_theme.

Tests added.
* REQUIRED_META added to test completeness.