boxes: Add a helper in MessageBox to update author status.
This commit adds a method to update author status of a specific message
box (if required).

The current logic re-builds the whole message box if an update is required.
This might need improvement as to update a specific element only. (Though
that might prove useful only if that element is rendered via a dedicated
method).

Tests added.