boxes: Implement syntax highlighting functionality.
This commit extracts the `code_soup` to be further parsed for css
styles. Each `codehilite` associated tag has a `pre` tag wrapped
around a `code` tag which contains `span` tags and NavigableStrings.
In some old messages, the `code` tag doesn't exist which is also
dealt with.

Basic syntax highlighting functionality was implemented in
soup2markup by extracting the `css_style` of each `span` tag in the
soup and using it as urwid style attributes. The remaining
NavigableStrings are given a style attribute of `w` which is the
whitespace css class in Pygments.

Tests amended.