test_core: Modify test_narrow_to_topic to use index_multiple_topic_msg.
This commit modifies the test_narrow_to_topic() function to leverage
the index_multiple_topic_msg fixture and test narrowing to a topic
that contains multiple messages. We add testing for cases when anchor
is provided to the _narrow_to function to enable us to broaden our
scope of testing and open our eyes to bugs that may have been unnoticed
with the previous test.

We also add asserts to check if the final focus after narrowing is on
the message we were expecting it to be in to ensure our code is robust.

Tests amended.