boxes/regexes: Add autocomplete helper for stream+topic prefix text.
This commit adds a method that would process and return typeaheads
for stream+topic prefixes. The current approach uses regex for
identifying the last occurence of required format. The dedicated
autocomplete function is: `autocomplete_stream_and_topic`.

Note that the regex used for accessing stream and topic data is
robust to the function, as the prefix would be manipulated to match
this regex type, and then a call to this function would be made. See
the next commit for more clarification.

The stream and topic names regexes are inspired from webapp's
static/js/compose_typeahead.js