core: Use linear scaling of popup size for small terminals.
This commit is a contribution towards #1005, allowing popups such as the
help menu to show properly on an 80-column width terminal.

This changes the fixed scaling factor of 3/4 to one that is linear in
the range 80 - 100 column width:
* The scaling fraction increases from 3/4 to 1 (full-width) as the
available size increases from 80 to 100. This increases the range of
sizes for which the popup is rendered properly.
* For width's smaller than 80 it defaults to full width.
* For width's greater than 100 it defaults to 3/4.

The values 80 and 100 are given by:
* MIN_SUPPORTED_POPUP_WIDTH = 80
* MAX_LINEAR_SCALING_WIDTH = 100

The popup height has a fixed scaling factor of 3/4, which retains
visibility of the title when the window is at the minimum supported
height.

Tests added.