bugfix: core: Fix message not focused in all_msg narrow if anchored.
This bugfix allows messages to be focused when narrowing to the
"all_messages" narrow from other narrows.

The primary bug was if the message from which we press 'a' was not
present in model.index['all_msg_ids'], then we were not fetching
that from the server. The reason we were not fetching those messages
was we only checked if the final narrow is not empty, which in this
case will never be, and not if it contains the message we are
focusing to. This made the focus to move elsewhere in the
"all_messages" narrow.

Related discussion: #zulip-terminal > Contextual narrowing bug.

Tests added.