refactor: conftest: Add factory function for generating recipients.
This commit adds a display_recipient_factory() in conftest that can
generate the "display_recipients" field in msg_template for message
belonging to either PM/group.

This function requires a list of tuples containing the recipient_id
and recipient_name for generating the fields. This function is not
defined as a fixture since we currently only need this to create
other fixtures and not directly in tests.