refactor: conftest: Generate msg templates from factory function.
This commit adds a new function msg_template_factory that can generate
templates for all message types, i.e. stream/PM/group with minimum
input parameters.

This new factory simplifies the code and reduces duplication in tests.
All the different message template functions are defined as fixtures
so that they can be reused throughout the tests. This new factory also
makes it easy to create a large number of output templates with minimal
modification of the input.