refactor: views: Relax the type annotation for PopUpViewTableContent.
This short commit updates the type annotation for the contents of
PopUpView. This is done so that there isn't a conflict in follow-up
commits where the table content would contain types such as
Tuple[str, Any]. Currently the tuple types supported are only
Tuple[str, str].

Note that `Any` here is not very explicit and clear, and so potentially
needs improvement. A FIXME is added noting the same.