bugfix: boxes: Allow message body editing when limit doesn't exist.
This commit fixes a bug that doesn't let users edit the message body
when the `realm_message_content_edit_limit_seconds` sent by the server
is 0 for a realm. Prior to this commit, this configuration of 0 was
treated as a complete restriction against message body editing, while
the configuration actually signified no limit in message body editing.

Fixes #1034.