tests: ui_tools: Include tests for footer text in EDIT_MESSAGE keypress.
This commit adds tests for footer notifications to make them clearer.
This allows it to be easily noticed and expanded if the footer texts are
changed in a particular commit.

We pass extra parameters to the test function, that holds the expected
footer texts in case of stream or private message editing, and assert
it with the value that the report_* functions have been called with.