refactor: tests: Simplify parametrize names & minimally indent.
This is similar to the previous commit, but also involves minimal indent
adjustment to pass linting.
(the minimal indenting will be adjusted automatically later upon
application of black)

The parametrize names being split into lists was originally introduced
due to breaking across lines, which is still required in some cases.