bugfix: api_types/model: Update mentioned narrow on message edit event.
This commit updates the index and unread_count if the message was indexed
when an update message event with change in mention flag(s) is detected.
If the message was not in index, we just update the unread_count. In this
case, the count is updated by comparing with initial_data's "unread_msgs"
field which does not stay synced with the server, hence the logic is buggy.

Changes in mentioned narrow after this event are not rendered at the same
time but it updates the model so that a return to the mentioned messages
narrow will do so.

The tests are also tried to be modified as less as possible, and the
present fixtures were mostly used to test for this new condition,
thus preventing the use of too many fixtures.

Tests amended.