bugfix: ui: Focus on compose area when draft is opened from side-panels.
This commit fixes a bug that caused the focus to not move to the central
area if the draft was opened from the side-panels. This is a minor bugfix
that helps redirect the focus to the compose area in such cases.

The recently added tests for OPEN_DRAFT are amended in this commit to
match the added bugfix.