DNM: SPECIAL COMMIT FOR REBASING AGAINST BLACK MIGRATION.
DO NOT MERGE THIS COMMIT.

This commit is a combined revert of the application of black to the
repo. It is intended for use with a small script to translate branches
into a black-compatible form.

BEFORE MIGRATING A BRANCH MAKE A COPY OF THE ORIGINAL IN ANOTHER BRANCH
- JUST IN CASE!

The migration process is:
* Fetch this branch, and rename it locally as `black-rebase`
* Checkout a branch to translate and rebase against `black-rebase`
  (this ensures it is up to date with the state before black)
* Ensure you have up to date dependencies, including black, using eg. `make`
* Use the following script to migrate your current branch so that each
  of your own commits is automatically converted:

FILTER_BRANCH_SQUELCH_WARNING=1 \
    git filter-branch --force \
    --tree-filter 'black zulipterminal/ tests/ --fast' \
    --commit-filter 'git_commit_non_empty_tree "$@"' \
    "black-rebase^..HEAD"