tests: Generalize the update message index using factory.
This commit adds a factory to generate the messages index relevant to
update message event tests, in an effort to reduce code duplications
for future tests. The factory requires three parameters - msg_stream_id,
(stream_id of the messages in the index), topics (dict containing streams
as keys, and a list of topics as the values, consistent with the structure
in the index) and message_ids (which is a list of the message ids to be
generated in the index). The factory uses these parameters to customize
the index.