refactor: tests: Extract mentioned messages as fixture for reusability.
This commit defines a fixture that return a combination of
mentioned message so that its easier to re-use it across various
tests that demands such test cases.

The index_all_mentioned fixture in conftest and the
test_index_mentioned_message in helper are also amended to use this
fixture instead of defining their own.

Tests amended.