boxes: Create helper function to update the stream compose header.
This commit adds a helper function facilitating abstracted updates
to the stream name in the stream compose header (title_write_box),
the corresponding edit_pos modification, as well as the stream_id
attribute of the WriteBox object in use. The helper also notifies
the user of the topic being moved to another stream using
`report_warning()`.

Test added.