model: Handle topic updates in the compose header mid-compose.
This commit enables dynamic behaviour in the compose box when a topic
is updated when the user is currently composing. The topic that the user
is currently writing a message to is checked and the relevant update is
done if the topic in the event matches the corresponding header.

Tests added.