model: Handle updating narrows depending on flags for delete_msg event.
This commit extends the previous commit that added rudimentary
support for handling delete_message event, and allows updating the
special narrows such as mentioned/starred that were depending on
message flags. This commit could be considered as a bugfix over the
previous one. It also handles updating the unread count if the
message was earlier unread before being deleted.

Tests added and amended.
Fixes one check-box in #993.